src/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

# Google Cloud
google-cloud-storage==2.16.0

# HTML & PDF manipulation
weasyprint==61.2
Jinja2==3.1.3

# utilities
surquest-fastapi-utils==0.2.6
surquest-utils-loader==0.0.1
surquest-utils-config-formatter==0.0.2